Enhancements
git-svn-id: https://emgucv.svn.sourceforge.net/svnroot/emgucv/trunk@139 d7f09016-e345-0410-b530-edf29a71df78

+++ Emgu.CV.Test/Class1.cs
@@ -145,10 +145,9 @@ namespace Emgu.CV.Test
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            using (Image<Bgr, Byte> img = new Image<Bgr, byte>("stuff.jpg"))
-            using (Image<Bgr, Byte> img2 = img.Resize(100, 100, true))
+            using (Image<Bgr, Single> img = new Image<Bgr, Single>("stuff.jpg"))
+            using (Image<Bgr, Single> img2 = img.Resize(100, 100, true))
             {
-
                 Application.Run(new ImageViewer(img2));
                 Rectangle<double> r = img2.ROI;
                 r.Size = new Point2D<double>(r.Size.X / 2, r.Size.Y / 2);
